Updated script

+++ scripts/clean_docker.sh
@@ -2,8 +2,4 @@
 
 docker volume prune --force
 docker network prune --force
-
-# Remove all unused image if `+i` option given
-if test "$1" = "+i"; then
-   docker image prune -a --force
-fi
+test "$1" = "-i" -o "$1" = "--prune-images" && docker image prune -a --force
